Auth/Manager

+++ clouds5spring/src/main/java/com/carte/clouds5spring/repository/RouteProblemeRepository.java
@@ -3,5 +3,5 @@ package com.carte.clouds5spring.repository;
 import com.carte.clouds5spring.entity.RouteProbleme;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RouteProblemeRepository extends JpaRepository<RouteProbleme, Long> {
+public interface RouteProblemeRepository extends JpaRepository<RouteProbleme, Integer> {
 }
